+++ src/Controller/ScenarioScheduler/ValueObject/ScenarioSchedulerListingItem.php
@@ -20,6 +20,8 @@ final class ScenarioSchedulerListingItem
 
     public DateTimeImmutable $updatedAt;
 
+    public bool $active;
+
     public string $expression;
 
     /** @var array<string, string> */
@@ -35,6 +37,7 @@ final class ScenarioSchedulerListingItem
         string $name,
         DateTimeImmutable $createdAt,
         DateTimeImmutable $updatedAt,
+        bool $active,
         string $expression,
         array $flags
     ) {
@@ -44,6 +47,7 @@ final class ScenarioSchedulerListingItem
         $this->name = $name;
         $this->createdAt = $createdAt;
         $this->updatedAt = $updatedAt;
+        $this->active = $active;
         $this->expression = $expression;
         $this->flags = $flags;
     }