+++ tests/Controller/ScenarioScheduler/requests/scenario-scheduler-request-body.array-config.php
@@ -9,6 +9,7 @@ return new ScenarioSchedulerRequestBody(
     [
         'optional' => 'true',
     ],
+    true,
     '0 1 * * *',
     [
         'options' => [