+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+## 1.2.0 - 2023-07-11
+### Added
+- Added method `ScenarioSchedulersController::activateScenarioScheduler()`.
+- Added method `ScenarioSchedulersController::deactivateScenarioScheduler()`.
+- Added property `bool $active` and mapping for value object `ScenarioSchedulerListingItem`, request body `ScenarioSchedulerRequestBody` and response body `ScenarioSchedulerResponseBody`.
+
+### Changed
+- Updated README.
+
 ## 1.1.1 - 2023-06-21
 ### Added
 - Added method `CrawlerClientInterface::getSerializer()`.